requirements.txt
django
djangorestframework
pillow
pytest-django